requirements.txt
requests
beautifulsoup4
flask
flask-restful
flask-restful-swagger
flask-sqlalchemy
